Granular tests for interaction with UCAS website.
The new integration tests in this patch test the interaction with the
live UCAS website. These tests are necessary as the generation of urls
for courses is complex and fragile to change on the UCAS side. These
tests will act as a canary for any changes to the UCAS site that we need
to react to and give an idea of what kind of change / failure we might
be dealing with.

The tests that are removed in this patch were too intertwined with
database, controllers etc. to give a precise view of the cause of any
failures.

Full integration of database + controllers + ucas + ui can be tested
with webdriver at the UI layer, ask Sau for his magic script :-)

The magic strings are now compiled into the library to give greater
certainty as to the behaviour of interaction with UCAS across all
environments. We can still respond quickly to changes as we are
sufficiently agile to be able to push changes to production rapidly.

* Refactored the UCAS-handling code out of the controller for cleaner
  testing.