Provider search prefers exact matches
We had an issue that a provider with a smaller number of courses
was drowned out by providers with more courses. Since they had a short name,
no text search could be formed to return this provider in the 5 suggestions.

This changes selection logic to say that if an exact match is available,
show this as the first result.