[1162] Fix postgres error for zero-distance location lookups
There seems to be a floating point rounding error in the haversine
formula that results in a distance >1 radians where the lat-lon in the
locations table exactly matches the location searched for.

This has been fixed by adding LEAST(x,1) to make sure the value doesn't
exceed the allowable range for ACOS.

Without this patch a search for Dorchester (which is in our location
table) resulted in postgres error "ERROR:  input is out of range".

We consider that given this is a UK based service we are not going to
fix the -1 rounding error case as that would be for searches on the
other side of the planet.

* Additionally reformatted the sql to make it easier to see the nesting
  of statements.

To see the actual diff:

kdiff3 20181204172354_FasterLocationQuery.cs 20190314145136_FixHaversineRoundingError.cs

More info:
https://stackoverflow.com/questions/2533386/postgresql-error-error-input-is-out-of-range/55166475#55166475