[1579] Fix dependency injection for geocoder
This was breaking the synchronise from publish to search and compare.
It seems there's been a change in the dotnet runtime that caused this to
start failing. LocationRequester was a singleton depending on the scoped
dbcontext service which actually doesn't make sense from a scoping point
of view.

> System.InvalidOperationException: Cannot consume scoped service
> 'GovUk.Education.SearchAndCompare.Api.DatabaseAccess.ICourseDbContext'
> from singleton
> 'GovUk.Education.SearchAndCompare.Geocoder.ILocationRequester'.

By making the geocoder use the same scope rule as the dbcontext they now
line up and this stops failing.