[1581] Add additional course info to circuit breaker
When the circuit breaker trips, we can see the number of added or
removed courses. It would be useful to also see the
providerCode/courseCode tuples of some courses that have been
added/removed.

I'm not sure what the ordering of `_context.Courses` or
`receivedCourses` is and whether this will produce an actual useful
output, but assuming that they are ordered by the `updated_at` or the
`id`, then this should work to output the course codes for the last 5
added or removed courses.